requirements.txt
fastapi
python-multipart
sqlalchemy
uvicorn
groq
python-dotenv
celery
redis
PyMuPDF